info only exception added

+++ Window.cpp
@@ -100,7 +100,7 @@ std::optional<int> Window::ProcessMessage() noexcept
 	{
 		if (msg.message == WM_QUIT)
 		{
-			return msg.wParam;
+			return (int)msg.wParam;
 		}
 
 		TranslateMessage(&msg);
